requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-groq
faiss-cpu
python-dotenv
pypdf
fastapi
uvicorn[standard]
python-multipart
mangum

